Updated four more DynamoDB examples.
