--- dotnetv3/dynamodb/CreateTablesLoadData/CreateTablesLoadData.csproj
@@ -1,13 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="AWSSDK.Core" Version="3.5.1.21" />
-    <PackageReference Include="AWSSDK.DynamoDBv2" Version="3.5.0.23" />
-  </ItemGroup>
-
-</Project>
