+++ dotnetv3/dynamodb/CreateTablesLoadDataExample/CreateTablesLoadDataExample.sln
@@ -1,0 +1,37 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30114.105
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CreateTablesLoadDataExample", "CreateTablesLoadDataExample\CreateTablesLoadDataExample.csproj", "{ED45F379-E94E-4245-99B8-FF61CB26250D}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{ED45F379-E94E-4245-99B8-FF61CB26250D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{ED45F379-E94E-4245-99B8-FF61CB26250D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{ED45F379-E94E-4245-99B8-FF61CB26250D}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{ED45F379-E94E-4245-99B8-FF61CB26250D}.Debug|x64.Build.0 = Debug|Any CPU
+		{ED45F379-E94E-4245-99B8-FF61CB26250D}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{ED45F379-E94E-4245-99B8-FF61CB26250D}.Debug|x86.Build.0 = Debug|Any CPU
+		{ED45F379-E94E-4245-99B8-FF61CB26250D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{ED45F379-E94E-4245-99B8-FF61CB26250D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{ED45F379-E94E-4245-99B8-FF61CB26250D}.Release|x64.ActiveCfg = Release|Any CPU
+		{ED45F379-E94E-4245-99B8-FF61CB26250D}.Release|x64.Build.0 = Release|Any CPU
+		{ED45F379-E94E-4245-99B8-FF61CB26250D}.Release|x86.ActiveCfg = Release|Any CPU
+		{ED45F379-E94E-4245-99B8-FF61CB26250D}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {B9109AE8-ED3B-4FCE-96BB-5C1B99B73DDD}
+	EndGlobalSection
+EndGlobal
