+++ dotnetv3/dynamodb/FromSQL/CreateIndexExample/CreateIndexExample.sln
@@ -1,0 +1,34 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30114.105
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CreateIndexExample", "CreateIndexExample\CreateIndexExample.csproj", "{6163B7B5-9C4D-4774-BB14-03CF5578FCF2}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{6163B7B5-9C4D-4774-BB14-03CF5578FCF2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6163B7B5-9C4D-4774-BB14-03CF5578FCF2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6163B7B5-9C4D-4774-BB14-03CF5578FCF2}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{6163B7B5-9C4D-4774-BB14-03CF5578FCF2}.Debug|x64.Build.0 = Debug|Any CPU
+		{6163B7B5-9C4D-4774-BB14-03CF5578FCF2}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{6163B7B5-9C4D-4774-BB14-03CF5578FCF2}.Debug|x86.Build.0 = Debug|Any CPU
+		{6163B7B5-9C4D-4774-BB14-03CF5578FCF2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6163B7B5-9C4D-4774-BB14-03CF5578FCF2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6163B7B5-9C4D-4774-BB14-03CF5578FCF2}.Release|x64.ActiveCfg = Release|Any CPU
+		{6163B7B5-9C4D-4774-BB14-03CF5578FCF2}.Release|x64.Build.0 = Release|Any CPU
+		{6163B7B5-9C4D-4774-BB14-03CF5578FCF2}.Release|x86.ActiveCfg = Release|Any CPU
+		{6163B7B5-9C4D-4774-BB14-03CF5578FCF2}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
