+++ dotnetv3/dynamodb/GetItemsExample/GetItemsExample.sln
@@ -1,0 +1,37 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30114.105
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GetItemsExample", "GetItemsExample\GetItemsExample.csproj", "{DE0899F5-FA16-4139-A3B7-FCEB449C0F11}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{DE0899F5-FA16-4139-A3B7-FCEB449C0F11}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DE0899F5-FA16-4139-A3B7-FCEB449C0F11}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DE0899F5-FA16-4139-A3B7-FCEB449C0F11}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{DE0899F5-FA16-4139-A3B7-FCEB449C0F11}.Debug|x64.Build.0 = Debug|Any CPU
+		{DE0899F5-FA16-4139-A3B7-FCEB449C0F11}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{DE0899F5-FA16-4139-A3B7-FCEB449C0F11}.Debug|x86.Build.0 = Debug|Any CPU
+		{DE0899F5-FA16-4139-A3B7-FCEB449C0F11}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DE0899F5-FA16-4139-A3B7-FCEB449C0F11}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DE0899F5-FA16-4139-A3B7-FCEB449C0F11}.Release|x64.ActiveCfg = Release|Any CPU
+		{DE0899F5-FA16-4139-A3B7-FCEB449C0F11}.Release|x64.Build.0 = Release|Any CPU
+		{DE0899F5-FA16-4139-A3B7-FCEB449C0F11}.Release|x86.ActiveCfg = Release|Any CPU
+		{DE0899F5-FA16-4139-A3B7-FCEB449C0F11}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {F2BFCC86-98CC-4243-A6FB-8F84BB8C3B91}
+	EndGlobalSection
+EndGlobal
