+++ dotnetv3/dynamodb/GetItemsExample/GetItemsExample/Properties/launchSettings.json
@@ -1,0 +1,8 @@
+{
+  "profiles": {
+    "GetItemsExample": {
+      "commandName": "Project",
+      "commandLineArgs": "-i 4"
+    }
+  }
+}
