+++ dotnetv3/dynamodb/GetItemsExample/app.config
